requirements.txt
streamlit==1.27.2
requests==2.31.0
Pillow==10.1.0
torch==2.1.0
fastapi==0.104.0
pydantic==2.4.2
torchvision==0.16.0


